fix(api): #60 code-review fixes + nodenext module resolution
High-effort 7-angle review of the #60 auth work; all gate-verified
(lint/test/build/db:check + rls-test.sh):

- guard: accept case-insensitive Bearer scheme (RFC 6750)
- auth: return 409 (not an unhandled 500) on the register email-uniqueness race
- sessions: ParseUUIDPipe on /auth/v1/sessions/:id (a malformed id hit Postgres as a 500)
- cookies: Secure gated by env so the cookie auth path works over local HTTP
- users: remove the unauthenticated /users HTTP controller (an email-enumeration
  oracle + an unhashed-password create); UsersService stays for internal DI

Architectural: switch apps/api to moduleResolution "nodenext" (+esModuleInterop)
so package `exports` maps resolve and CJS/ESM mismatches fail at build, not at
runtime. Replace bcrypt-ts (ESM-only types — unimportable at the top of a CJS
module) with bcryptjs, which ships CommonJS types under its `require` condition,
so password hashing is a normal top-level import — no path hack, no dynamic import.
