+++ apps/api/package.json
@@ -36,7 +36,7 @@
     "@nestjs/platform-express": "^11.0.1",
     "@nestjs/swagger": "^11.4.4",
     "ai": "5.0.0-beta.12",
-    "bcrypt-ts": "^7.1.0",
+    "bcryptjs": "^3.0.3",
     "class-transformer": "^0.5.1",
     "class-validator": "^0.15.1",
     "dotenv": "^17.0.0",
